requirements.txt
textual
httpx
PyYAML
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
